requirements.txt
python-chess
pandas
openai
google-genai